Maybe try code

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -1,17 +1,21 @@
 package ru.kata.spring.boot_security.demo.service;
 
-import ru.kata.spring.boot_security.demo.models.Role;
 import ru.kata.spring.boot_security.demo.models.User;
-
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 public interface UserService {
+
     void add(User user);
-    void update(Long id, User user);
-    void delete (Long id);
+
+    void update(int id, User user);
+
+    void delete(int id);
+
     List<User> findAll();
-    User findById(Long id);
+
+    User findById(int id);
+
     Optional<User> findByEmail(String email);
+
 }
